gcc/riscv: Include more registers in SIBCALL_REGS
The current SIBCALL_REGS are x6, x7, and x28 to x31.  These are all
caller saved registers, however, they are not all of the caller saved
registers.

I don't see any reason why we couldn't add t1, and a0 to a7 into this
set, and this is what this patch does.

gcc/ChangeLog:

	* config/riscv/riscv.h (REG_CLASS_CONTENTS): Update SIBCALL_REGS.