Replace APB_DECR environment variable with a flag
The flag:

  -finline-growth-bias=<number>

can now be used to control the bias of the inline growth, instead
of setting the APB_DECR environment variable. Note that the value
is inverted compared to APB_DECR - e.g. -finline-growth-bias=-2 is
equivalent to APB_DECR=2. The rationale behind this is to make the
functionality more comprehensible when posting the patch on the
GCC mailing list - otherwise, to bias the growth positively, a
negative value is required, which is confusing.

There is not a great deal of validation of the argument - if a
non-integer is passed as the parameter (e.g. -finline-growth=xyz),
it will be as if the parameter were 0. More validation could be
added if necessary. GCC's options infrastructure doesn't seem to
anticipate option values / parameters that could contain negative
values. For parameters, -1 represents an error and results in an
ICE (-2 etc. pass through OK though). For options, only the
UInteger and Host_Wide_Int numeric types are provided, which both
expect a positive integer. I did consider extending this with an
Integer type that could accept positive and negative integers,
(e.g. starting with augmenting switch_bit_fields in
opt-functions.awk) but rooting out assumptions of non-negative
integer values in places further down seemed like a time-consuming
endeavour. In conclusion, validation of the argument should
probably be done by adding a custom check to the big switch in
common_handle_option in opts.c.